+++ tests/test_character_ui.py
@@ -48,7 +48,7 @@ const elements = {
   'character-slots-left': new Elem('div'),
   'character-slots-right': new Elem('div'),
   'equipment-items': new Elem('div'),
-  'character-image': new Elem('div'),
+  'character-image': new Elem('div')
 };
 
 global.document = {